+++ CHANGELOG.md
@@ -6,11 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
+## [3.0.1] - 2025-06-23
+
+### Changed
+- Better naming of string escaping functions and constants in result writer.
+
 ## [3.0.0] - 2025-06-23
 
 ### Changed
 - Output is now one precisely defined text file format. CSV files are no longer supported as Excel has strange and bizarre conversion rules.
-- The option "separator" is therefore no longer present. 
+- The option "separator" is therefore no longer present.
 
 ## [2.5.0] - 2025-06-22